there is a setting to allow a choice between HOSTNAME or MX check for sender-verification the settings panel showed the option to use the default-sender in case this is a verified sender, but saving the value reverted the choice because during saving the settings, the choice for MX was not respected
This choice was also not respected during sending. The option panel was therefore not functional.

These bugs have been fixed, the e-mails sent now have the correct sender, if MX verification is chosen and 
the default sender is not on the same domain as the website and that option has been chosen to use as new
sender in case the sender sending the email is not valid.

Jeez, I have trouble typing it, so I don't blame you if you need to read it more than twice

git-svn-id: svn://svn.clearsite.nl/wp_plugins/wp-email-essentials/trunk@8035 ad331a6c-ddf7-4250-94be-11eacc3bd02f